Compiling with --disalbe-stack-for-recursion is working. Introduce PCRE2_OFFSET for offset values.

+++ src/pcre2_printint.c
@@ -282,7 +282,7 @@ pcre2_printint(pcre2_code *re, FILE *f, BOOL print_lengths)
 {
 PCRE2_SPTR codestart, nametable, code;
 uint32_t options = re->compile_options;
-size_t nesize = re->name_entry_size;
+uint32_t nesize = re->name_entry_size;
 BOOL utf = (options & PCRE2_UTF) != 0;
 
 nametable = (PCRE2_SPTR)((uint8_t *)re + sizeof(pcre2_real_code));
